# TST: sparse: Separate Test64Bit from test_base.py, refactor to allow Test_Classes based marks, ignore some methods