--- a/scipy/sparse/tests/test_64bit.py
+++ b/scipy/sparse/tests/test_64bit.py
@@ -25,8 +25,6 @@ from .test_base import (
     with_64bit_maxval_limit,
 )
 
-counter_container = [0]
-zero_get_index_dtype_calls : set[str] = set()
 
 # name : reason not tested here
 SKIP_TESTS = {
@@ -39,11 +37,10 @@ SKIP_TESTS = {
     'test_constructor_largecol': 'test verifies int64 indexes',
     'test_tocoo_tocsr_tocsc_gh19245': 'test verifies int32 indexes',
 }
-# See the bottom of the module for IGNORE_TESTS
 
 
-def cases_64bit(sp_api, ignore=True):
-    """Yield all tests for all formats that use get_index_dtype
+def cases_64bit(sp_api):
+    """Yield all tests for all formats
 
     This is more than testing get_index_dtype. It allows checking whether upcasting
     or downcasting the index dtypes affects test results. The approach used here
@@ -51,7 +48,6 @@ def cases_64bit(sp_api, ignore=True):
     We just run them all.
     So, each test method in that uses cases_64bit reruns most of the test suite!
     """
-    IGNORE = IGNORE_TESTS if ignore else {}
     if sp_api == "sparray":
         TEST_CLASSES = [_TestBSR, _TestCOO, _TestCSC, _TestCSR, _TestDIA]
     elif sp_api == "sparray-extra":
@@ -77,9 +73,6 @@ def cases_64bit(sp_api, ignore=True):
                     not getattr(method, 'slow', False)):
                 marks = []
 
-                if name in IGNORE:
-                    continue
-
                 msg = SKIP_TESTS.get(method_name)
                 if msg:
                     marks.append(pytest.mark.skip(reason=msg))
@@ -92,6 +85,7 @@ def cases_64bit(sp_api, ignore=True):
                 yield pytest.param(cls, method_name, marks=marks)
 
 
+@pytest.mark.thread_unsafe
 class RunAll64Bit:
     def _check_resiliency(self, cls, method_name, **kw):
         # Resiliency test, to check that sparse matrices deal reasonably
@@ -110,29 +104,6 @@ class RunAll64Bit:
 
         check(cls, method_name)
 
-    def _check_resiliency_and_list_of_ignores(self, cls, method_name):
-        @with_64bit_maxval_limit(random=counter_container)
-        def check_and_collect(cls, method_name):
-            instance = cls()
-            if hasattr(instance, 'setup_method'):
-                instance.setup_method()
-            try:
-                counter_container[0] = 0
-                getattr(instance, method_name)()
-                if counter_container[0] == 0:
-                    zero_calls.add(name)
-            finally:
-                if hasattr(instance, 'teardown_method'):
-                    instance.teardown_method()
-
-        name = f'{cls.__name__}-{method_name}'
-        zero_calls = zero_get_index_dtype_calls
-
-        check_and_collect(cls, method_name)
-
-        if name in zero_calls:
-            assert name in IGNORE_TESTS
-
 
 class Test64BitArray(RunAll64Bit):
     # inheritance of pytest test classes does not separate marks for subclasses.
@@ -150,9 +121,9 @@ class Test64BitArray(RunAll64Bit):
         self._check_resiliency(cls, method_name, fixed_dtype=np.int64)
 
     @pytest.mark.fail_slow(2)
-    @pytest.mark.parametrize('cls,method_name', cases_64bit("sparray", ignore=False))
+    @pytest.mark.parametrize('cls,method_name', cases_64bit("sparray"))
     def test_resiliency_random(self, cls, method_name):
-        self._check_resiliency_and_list_of_ignores(cls, method_name)
+        self._check_resiliency(cls, method_name)
 
 
 class Test64BitMatrix(RunAll64Bit):
@@ -180,10 +151,10 @@ class Test64BitMatrixSameAsArray(RunAll64Bit):
         self._check_resiliency(cls, method_name, fixed_dtype=np.int64)
 
     @pytest.mark.fail_slow(2)
-    @pytest.mark.parametrize('cls,method_name', cases_64bit("spmatrix", ignore=False))
+    @pytest.mark.parametrize('cls,method_name', cases_64bit("spmatrix"))
     def test_resiliency_random(self, cls, method_name):
         # Resiliency check that sparse deals with varying index data types.
-        self._check_resiliency_and_list_of_ignores(cls, method_name)
+        self._check_resiliency(cls, method_name)
 
 
 @pytest.mark.xslow
@@ -203,10 +174,10 @@ class Test64BitArrayExtra(RunAll64Bit):
         self._check_resiliency(cls, method_name, fixed_dtype=np.int64)
 
     @pytest.mark.fail_slow(2)
-    @pytest.mark.parametrize('cls,method_name', cases_64bit("sparray-extra", False))
+    @pytest.mark.parametrize('cls,method_name', cases_64bit("sparray-extra"))
     def test_resiliency_random(self, cls, method_name):
         # Resiliency check that sparse deals with varying index data types.
-        self._check_resiliency_and_list_of_ignores(cls, method_name)
+        self._check_resiliency(cls, method_name)
 
 
 @pytest.mark.xslow
@@ -232,26 +203,13 @@ class Test64BitMatrixExtra(RunAll64Bit):
         self._check_resiliency(cls, method_name, fixed_dtype=np.int64)
 
     @pytest.mark.fail_slow(2)
-    @pytest.mark.parametrize('cls,method_name', cases_64bit("spmatrix-extra", False))
+    @pytest.mark.parametrize('cls,method_name', cases_64bit("spmatrix-extra"))
     def test_resiliency_random(self, cls, method_name):
         # Resiliency check that sparse deals with varying index data types.
-        self._check_resiliency_and_list_of_ignores(cls, method_name)
+        self._check_resiliency(cls, method_name)
 
 
-@pytest.mark.xslow
-def test_report_zero_get_index_dtype_calls():
-    print(zero_get_index_dtype_calls)
-    print('IGNORE_TESTS = {')
-    for name in sorted(zero_get_index_dtype_calls):
-        print(f"    '{name}': 'get_index_dtype not called',")
-    print('}')
-
-    # these < unless you run entire module with xslow and slow. Then ==
-    # Test order matters here. Must be after all Test64Bit___.test_resiliency_random
-    # Usually put last in the order (see conftest.py)
-    assert zero_get_index_dtype_calls == IGNORE_TESTS.keys(), 'Found New Tests'
-
-
+@pytest.mark.thread_unsafe
 class Test64BitTools:
     # classes that use get_index_dtype
     MAT_CLASSES = [
@@ -344,88 +302,3 @@ class Test64BitTools:
         check_unlimited(csc_array, csr_array, coo_array)
         check_limited(csc_matrix, csr_matrix, coo_matrix)
         check_unlimited(csc_matrix, csr_matrix, coo_matrix)
-
-
-# We could track includes rather than ignores. But this way no new tests get skipped
-IGNORE_TESTS = {
-#}
-#BYPASS = {
-    'TestBSR-test_invalid_shapes': 'get_index_dtype not called',
-    'TestBSR-test_rmul_scalar_type_error': 'get_index_dtype not called',
-    'TestBSRMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCOO-test_add_dense': 'get_index_dtype not called',
-    'TestCOO-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCOO-test_radd': 'get_index_dtype not called',
-    'TestCOOMatrix-test_add_dense': 'get_index_dtype not called',
-    'TestCOOMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCOOMatrix-test_radd': 'get_index_dtype not called',
-    'TestCSC-test_add_dense': 'get_index_dtype not called',
-    'TestCSC-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCSC-test_radd': 'get_index_dtype not called',
-    'TestCSC-test_rmul_scalar_type_error': 'get_index_dtype not called',
-    'TestCSCMatrix-test_add_dense': 'get_index_dtype not called',
-    'TestCSCMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCSCMatrix-test_radd': 'get_index_dtype not called',
-    'TestCSR-test_add_dense': 'get_index_dtype not called',
-    'TestCSR-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCSR-test_radd': 'get_index_dtype not called',
-    'TestCSR-test_rmul_scalar_type_error': 'get_index_dtype not called',
-    'TestCSRMatrix-test_add_dense': 'get_index_dtype not called',
-    'TestCSRMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestCSRMatrix-test_radd': 'get_index_dtype not called',
-    'TestDIA-test_invalid_shapes': 'get_index_dtype not called',
-    'TestDIA-test_setdiag_dtype': 'get_index_dtype not called',
-    'TestDIAMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestDIAMatrix-test_setdiag_dtype': 'get_index_dtype not called',
-    'TestDOK-test_dtype_preservation_empty_index': 'get_index_dtype not called',
-    'TestDOK-test_fancy_assignment_dtypes': 'get_index_dtype not called',
-    'TestDOK-test_invalid_shapes': 'get_index_dtype not called',
-    'TestDOK-test_negative_index_assignment': 'get_index_dtype not called',
-    'TestDOK-test_scalar_assign_2': 'get_index_dtype not called',
-    'TestDOK-test_ticket1160': 'get_index_dtype not called',
-    'TestDOKMatrix-test_dtype_preservation': 'get_index_dtype not called',
-    'TestDOKMatrix-test_dtype_preservation_empty_index': 'get_index_dtype not called',
-    'TestDOKMatrix-test_dtype_preservation_empty_slice': 'get_index_dtype not called',
-    'TestDOKMatrix-test_fancy_assignment_dtypes': 'get_index_dtype not called',
-    'TestDOKMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestDOKMatrix-test_negative_index_assignment': 'get_index_dtype not called',
-    'TestDOKMatrix-test_rmul_scalar_type_error': 'get_index_dtype not called',
-    'TestDOKMatrix-test_scalar_assign_2': 'get_index_dtype not called',
-    'TestDOKMatrix-test_ticket1160': 'get_index_dtype not called',
-    'TestLIL-test_copy': 'get_index_dtype not called',
-    'TestLIL-test_dtype_preservation_empty_index': 'get_index_dtype not called',
-    'TestLIL-test_empty': 'get_index_dtype not called',
-    'TestLIL-test_fancy_indexing_multidim_set': 'get_index_dtype not called',
-    'TestLIL-test_fancy_indexing_set': 'get_index_dtype not called',
-    'TestLIL-test_idiv_scalar': 'get_index_dtype not called',
-    'TestLIL-test_imul_scalar': 'get_index_dtype not called',
-    'TestLIL-test_index_scalar_assign': 'get_index_dtype not called',
-    'TestLIL-test_invalid_shapes': 'get_index_dtype not called',
-    'TestLIL-test_negative_index_assignment': 'get_index_dtype not called',
-    'TestLIL-test_pickle': 'get_index_dtype not called',
-    'TestLIL-test_scalar_mul': 'get_index_dtype not called',
-    'TestLIL-test_setdiag_comprehensive': 'get_index_dtype not called',
-    'TestLIL-test_setelement': 'get_index_dtype not called',
-    'TestLIL-test_slice_assign_2': 'get_index_dtype not called',
-    'TestLIL-test_slice_scalar_assign': 'get_index_dtype not called',
-    'TestLIL-test_truediv_scalar': 'get_index_dtype not called',
-    'TestLILMatrix-test_copy': 'get_index_dtype not called',
-    'TestLILMatrix-test_dtype_preservation': 'get_index_dtype not called',
-    'TestLILMatrix-test_dtype_preservation_empty_index': 'get_index_dtype not called',
-    'TestLILMatrix-test_dtype_preservation_empty_slice': 'get_index_dtype not called',
-    'TestLILMatrix-test_empty': 'get_index_dtype not called',
-    'TestLILMatrix-test_fancy_indexing_multidim_set': 'get_index_dtype not called',
-    'TestLILMatrix-test_fancy_indexing_set': 'get_index_dtype not called',
-    'TestLILMatrix-test_idiv_scalar': 'get_index_dtype not called',
-    'TestLILMatrix-test_imul_scalar': 'get_index_dtype not called',
-    'TestLILMatrix-test_index_scalar_assign': 'get_index_dtype not called',
-    'TestLILMatrix-test_invalid_shapes': 'get_index_dtype not called',
-    'TestLILMatrix-test_negative_index_assignment': 'get_index_dtype not called',
-    'TestLILMatrix-test_pickle': 'get_index_dtype not called',
-    'TestLILMatrix-test_scalar_mul': 'get_index_dtype not called',
-    'TestLILMatrix-test_setdiag_comprehensive': 'get_index_dtype not called',
-    'TestLILMatrix-test_setelement': 'get_index_dtype not called',
-    'TestLILMatrix-test_slice_assign_2': 'get_index_dtype not called',
-    'TestLILMatrix-test_slice_scalar_assign': 'get_index_dtype not called',
-    'TestLILMatrix-test_truediv_scalar': 'get_index_dtype not called',
-}