--- a/scipy/sparse/tests/test_base.py
+++ b/scipy/sparse/tests/test_base.py
@@ -4367,11 +4367,11 @@ class TestCSRMatrix(_MatrixMixin, TestCSR):
             sup.filter(SparseEfficiencyWarning, "Changing the sparsity structure")
             return csr_matrix(*args, **kwargs)
     
-    def test_spmatrix_subscriptable(self):
-        result = csr_matrix[np.int8]
-        assert isinstance(result, GenericAlias)
-        assert result.__origin__ is csr_matrix
-        assert result.__args__ == (np.int8,)
+def test_spmatrix_subscriptable():
+    result = csr_matrix[np.int8]
+    assert isinstance(result, GenericAlias)
+    assert result.__origin__ is csr_matrix
+    assert result.__args__ == (np.int8,)
 
 
 TestCSRMatrix.init_class()
@@ -4908,24 +4908,24 @@ class TestCOO(sparse_test_class(getset=False,
         assert_((mat1.reshape((1001, 3000001), order='C') != mat2).nnz == 0)
         assert_((mat2.reshape((3000001, 1001), order='F') != mat1).nnz == 0)
     
-    def test_sparray_subscriptable(self):
-        result = coo_array[np.int8, tuple[int]]
-        assert isinstance(result, GenericAlias)
-        assert result.__origin__ is coo_array
-        assert result.__args__ == (np.int8, tuple[int])
-
-        result = coo_array[np.int8]
-        assert isinstance(result, GenericAlias)
-        assert result.__origin__ is coo_array
-        assert result.__args__ == (np.int8,)
-
-
 class TestCOOMatrix(_MatrixMixin, TestCOO):
     spcreator = coo_matrix
 
 
 TestCOO.init_class()
 TestCOOMatrix.init_class()
+
+
+def test_sparray_subscriptable():
+    result = coo_array[np.int8, tuple[int]]
+    assert isinstance(result, GenericAlias)
+    assert result.__origin__ is coo_array
+    assert result.__args__ == (np.int8, tuple[int])
+
+    result = coo_array[np.int8]
+    assert isinstance(result, GenericAlias)
+    assert result.__origin__ is coo_array
+    assert result.__args__ == (np.int8,)
 
 
 class TestDIA(sparse_test_class(getset=False, slicing=False, slicing_assign=False,