This PR is a followup/replacement for #21560 and implements some ideas from that discussion.

The goal is to allow essentially the same test coverage with reduced resources. Each of the following features can be removed if desired without affecting the PR too much.
The PR
1) refactors the Test64Bit classes to allow DOK and LIL to be marked xslow
2) splits the Test64Bit suite of tests out of test_base.py into its own module (eases partial testing)
3) ignores: a hardcoded IGNORE_TESTS dict of 78 tests that don't call `get_index_dtype`. No 64bit issues.

**I'd like feedback**:
- which categories should be marked as slow or xslow?
- Should I take out the code to ignore methods that dont use `get_index_dtype`?

**Notation**:
"Array" and "Matrix" formats: `BSR, COO, CSC, CSR, DIA` all use index dtypes.
"Extra": Formats `DOK, LIL` dont use index dtypes, but do convert to the other formats.

"Std": the 4 standard methods: `test_resiliency_*`: `limit_10`, `all_32`, `all_64` and `random`
"No64": the method to test down-casting that is only relevant for spmatrix. `test_no_64`

How frequently? (could also choose to drop completely)
"Always": runs 12 times per push (5 linux, 1 musllinux, 1 linux_OneAPI, 3 macos, 2 windows) 
"slow": runs 3 times per push (1 linux, 1 32bit linux, 1 windows)
    For "slow" I have included code in the 32-bit Linux workflow to run the slow tests even when mode is "fast".
"xslow": runs occasionally, not every push (every month or few months or so?)

**Current State**: We currently run Matrix with Extra in "fast" while Array with Extra is marked "slow". No test methods are ignored. Turning on sparray increases time from 45s to 65s (for each environment of the tests). 

This PR proposes to run Array and the No64 Matrix method in "fast", the other Matrix methods in "slow" and both Extra (Array and Matrix) as `xslow`.  Note that "slow" still gets run 3 times for each push (I have set up the 32-bit run to run slow Test64Bit tests.) This PR also sets up ignoring (not rerunning) test methods that don't call `get_index_dtype`.

**Proposal:**

|       | Array |Matrix |ArrayExtra |MatrixExtra|
| ----- | ----- | ----- | --------- | --------- |
|**Always** | Std   |no64   |           |           |    
|**slow**  |       |Std    |           |           |
|**xslow**  |       |       |Std        |Std & no64 |  

We end up ignoring 78 out of 2131 rerun tests within each Test64Bit test method.
That saves a couple seconds off each local run of the test suite.

**Timing results:**
```
main All Tests          65s   10809 tests
main mode fast          45s    7307 tests (Test64BitArray is marked slow)

 PR  All tests          63s   10536 tests (xslow, slow, standard...)
 PR  mode full          54s    8229 tests (slow, standard)
 PR  mode fast          36s    5795 tests

 PR  remove ignore code 38s    5852 tests (standard w/o ignoring no-call tests)
```

Details for Ignore: the code to track ignored test methods is a a little clumsy.
The tests use a hardcoded dict `IGNORE_TESTS` to look up test class/name for
any that should be skipped. To construct this hardcoded dict, we would run a
specific test occasionally to look for new tests that don't use `get_index_dtype`.
That test uses existing monkey patching machinery to increment a count on
each call to `get_index_dtype`, and stores the result across tests by using a
module scope list. Any test method that doesn't call is added to a dict.
When done, a that dict is compares with `IGNORE_TESTS` and if different,
a new hardcoded dict is printed and assertion is raised. That processing
test is marked "xslow".  Also, the final check must wait until after all the tests
run so conftest.py (in this PR) has a hook to put that test last.

**Summary:**
1) **refactors**: Seems like a good tradeoff and is in line with [gh-21560:comment](https://github.com/scipy/scipy/pull/21560#pullrequestreview-2339600169).
2) **module split**: Running test_base.py locally now takes ~half the time it used to. Better for local testing.
3) **ignores**: The 2s savings isn't much, and there is cost in terms of maintaining code, making sure the order of tests works, maintaining list of ignored tests. But there's no cost in terms of testing. Those tests are already run in the test_base classes.