@dschult 2025-04-29
We didn't need the workflow job to ensure spmatrix runs on 32bit because we aren't reducing the spmatrix testing (when I put that in, I was thinking we might make spmatrix run as "slow" to switch sparray to "fast".  But we've put both on "fast" now.  So that clean-up, the `name` not used clean-up and rebasing will hopefully lead to green checks here! (free-threading(1) is sometimes finicky)