@rgommers 2025-04-29
> `RuntimeError: cannot enter suppress_warnings twice.`

You removed all the `@pytest.mark.thread_unsafe` decorators, which explains those failures. 

Note that this isn't too hard to reproduce locally, building under free-threaded Python should work as expected when you use Cython >=3.1.0b1. Or perhaps even easier, follow https://github.com/rgommers/pixi-dev-scipystack/?tab=readme-ov-file#getting-started and run `pixi run test-nogil -s sparse -- --parallel-threads=2`.