@dschult 2025-04-29
Thanks @rgommers !!  
I've got it working locally and can reproduce the CI results. 
I think I didn't remove the decorators -- this PR is from before they were added... :)  
Should be an easy fix.