@dschult 2025-05-01
Looks like using `git diff  main -- scipy/sparse`
before `git rebase -i main` would have helped me avoid missing these changes with free-threading tests. 
Added to my toolkit -- and I think this is ready.